fix (#5): the tests package should not be renamed
