+++ packages/apis/backup_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/labstack/echo/v5"
 	"github.com/pocketbase/pocketbase/packages/core"
-	"github.com/pocketbase/pocketbase/packages/tests"
+	"github.com/pocketbase/pocketbase/tests"
 	"gocloud.dev/blob"
 )
