+++ packages/apis/base_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/labstack/echo/v5"
 	"github.com/pocketbase/pocketbase/packages/apis"
-	"github.com/pocketbase/pocketbase/packages/tests"
+	"github.com/pocketbase/pocketbase/tests"
 	"github.com/spf13/cast"
 )
