+++ packages/apis/file_test.go
@@ -11,8 +11,8 @@ import (
 	"github.com/labstack/echo/v5"
 	"github.com/pocketbase/pocketbase/packages/core"
 	"github.com/pocketbase/pocketbase/packages/daos"
-	"github.com/pocketbase/pocketbase/packages/tests"
 	"github.com/pocketbase/pocketbase/packages/tools/types"
+	"github.com/pocketbase/pocketbase/tests"
 )
 
 func TestFileToken(t *testing.T) {
