+++ packages/core/base_settings_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 
 	"github.com/pocketbase/pocketbase/packages/models"
-	"github.com/pocketbase/pocketbase/packages/tests"
 	"github.com/pocketbase/pocketbase/packages/tools/types"
+	"github.com/pocketbase/pocketbase/tests"
 )
 
 func TestBaseAppRefreshSettings(t *testing.T) {
