+++ packages/daos/admin_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/pocketbase/pocketbase/packages/models"
-	"github.com/pocketbase/pocketbase/packages/tests"
+	"github.com/pocketbase/pocketbase/tests"
 )
 
 func TestAdminQuery(t *testing.T) {
