+++ packages/daos/base_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/pocketbase/pocketbase/packages/daos"
 	"github.com/pocketbase/pocketbase/packages/models"
-	"github.com/pocketbase/pocketbase/packages/tests"
+	"github.com/pocketbase/pocketbase/tests"
 )
 
 func TestNew(t *testing.T) {
