+++ packages/daos/collection_test.go
@@ -10,9 +10,9 @@ import (
 	"github.com/pocketbase/pocketbase/packages/daos"
 	"github.com/pocketbase/pocketbase/packages/models"
 	"github.com/pocketbase/pocketbase/packages/models/schema"
-	"github.com/pocketbase/pocketbase/packages/tests"
 	"github.com/pocketbase/pocketbase/packages/tools/list"
 	"github.com/pocketbase/pocketbase/packages/tools/types"
+	"github.com/pocketbase/pocketbase/tests"
 )
 
 func TestCollectionQuery(t *testing.T) {
