+++ packages/forms/admin_password_reset_confirm_test.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/pocketbase/pocketbase/packages/forms"
 	"github.com/pocketbase/pocketbase/packages/models"
-	"github.com/pocketbase/pocketbase/packages/tests"
 	"github.com/pocketbase/pocketbase/packages/tools/security"
+	"github.com/pocketbase/pocketbase/tests"
 )
 
 func TestAdminPasswordResetConfirmValidateAndSubmit(t *testing.T) {
