+++ packages/forms/admin_upsert_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/pocketbase/pocketbase/packages/forms"
 	"github.com/pocketbase/pocketbase/packages/models"
-	"github.com/pocketbase/pocketbase/packages/tests"
+	"github.com/pocketbase/pocketbase/tests"
 )
 
 func TestNewAdminUpsert(t *testing.T) {
