+++ packages/tokens/admin_test.go
@@ -3,8 +3,8 @@ package tokens_test
 import (
 	"testing"
 
-	"github.com/pocketbase/pocketbase/packages/tests"
 	"github.com/pocketbase/pocketbase/packages/tokens"
+	"github.com/pocketbase/pocketbase/tests"
 )
 
 func TestNewAdminAuthToken(t *testing.T) {
