+++ packages/tools/filesystem/file_test.go
@@ -8,8 +8,8 @@ import (
 	"testing"
 
 	"github.com/labstack/echo/v5"
-	"github.com/pocketbase/pocketbase/packages/tests"
 	"github.com/pocketbase/pocketbase/packages/tools/filesystem"
+	"github.com/pocketbase/pocketbase/tests"
 )
 
 func TestNewFileFromPath(t *testing.T) {
